refactor(apps): extract Nostr sandbox bridge helpers to their own file
The Android attestation work pushed nostr_app_sandbox_screen.dart past
its frozen 842-line file-size ceiling (epic #4339), failing the
Generated Files / Mobile CI gate on #5318. Move the pure bridge helpers
— buildBridgeBootstrapScript, injectBridgeBootstrapIntoHtml,
webMessageAllowedOriginRules, and the private _escapeJs — into a focused
nostr_app_sandbox_bridge.dart, dropping the screen to 718 lines. The
_SandboxStatusCard widget and the _BootstrappedSandboxPage data holder
stay with the screen (still its only consumers).

Drop @visibleForTesting from the three helpers that now have
cross-library production callsites (the annotation would otherwise be an
analyzer error). Document the best-effort auto-login JS no-op so the
textual empty-catch detector stops flagging the bootstrap string. Shrink
the file_sizes and empty_catches baselines to lock in the wins.
