+++ mobile/lib/screens/apps/nostr_app_sandbox_screen.dart
@@ -15,6 +15,7 @@ import 'package:http/http.dart' as http;
 import 'package:nostr_app_bridge_repository/nostr_app_bridge_repository.dart';
 import 'package:openvine/l10n/l10n.dart';
 import 'package:openvine/providers/app_providers.dart';
+import 'package:openvine/screens/apps/nostr_app_sandbox_bridge.dart';
 import 'package:openvine/widgets/apps/nostr_app_permission_prompt_sheet.dart';
 import 'package:unified_logger/unified_logger.dart';
 import 'package:webview_flutter/webview_flutter.dart';
@@ -645,160 +646,6 @@ class _NostrAppSandboxScreenState extends ConsumerState<NostrAppSandboxScreen> {
   }
 }
 
-/// Builds the bridge bootstrap JavaScript with optional eager pubkey,
-/// provider metadata, ready-event dispatch, and per-app auto-login.
-///
-/// [nonce] is a per-mount secret embedded in every outgoing bridge
-/// request. The host rejects messages whose nonce does not match,
-/// which prevents iframes that bypass `window.nostr` and call the
-/// `divineSandboxBridge` channel directly from impersonating the main
-/// frame.
-@visibleForTesting
-String buildBridgeBootstrapScript({
-  required String nonce,
-  String? pubkey,
-  String? autoLoginScript,
-}) {
-  final escapedPubkey = pubkey != null && pubkey.isNotEmpty
-      ? _escapeJs(pubkey)
-      : '';
-  final escapedNonce = _escapeJs(nonce);
-  final loginJs = autoLoginScript != null && autoLoginScript.isNotEmpty
-      ? autoLoginScript.replaceAll('{{PUBKEY}}', escapedPubkey)
-      : '';
-
-  return '''
-(() => {
-  if (window.__divineNostrBridgeInstalled) {
-    return;
-  }
-
-  // Refuse to install in non-main frames. Cross-origin access to
-  // window.top throws; treat that as "not main frame" and bail.
-  try {
-    if (window.top !== window.self) {
-      return;
-    }
-  } catch (_) {
-    return;
-  }
-
-  const __divineBridgeNonce = '$escapedNonce';
-  const pending = new Map();
-  let nextId = 0;
-
-  const request = (method, args) => {
-    const id = `divine-\${++nextId}`;
-    const payload = JSON.stringify({
-      id,
-      method,
-      args: args ?? {},
-      nonce: __divineBridgeNonce,
-    });
-
-    return new Promise((resolve, reject) => {
-      pending.set(id, { resolve, reject });
-      divineSandboxBridge.postMessage(payload);
-    });
-  };
-
-  window.__divineNostrBridge = {
-    handleResponse(response) {
-      const pendingRequest = pending.get(response.id);
-      if (!pendingRequest) {
-        return;
-      }
-
-      pending.delete(response.id);
-
-      if (response.success) {
-        pendingRequest.resolve(response.result);
-        return;
-      }
-
-      const error = response.error || { code: 'bridge_error' };
-      const exception = new Error(error.message || error.code);
-      exception.code = error.code;
-      pendingRequest.reject(exception);
-    },
-  };
-
-  window.nostr = {
-    _pubkey: '$escapedPubkey' || null,
-    _metadata: {
-      name: 'diVine',
-      version: '1.0',
-      supports: ['nip04', 'nip44'],
-    },
-    getPublicKey() {
-      if (this._pubkey) return Promise.resolve(this._pubkey);
-      return request('getPublicKey', {});
-    },
-    getRelays() {
-      return request('getRelays', {});
-    },
-    signEvent(event) {
-      return request('signEvent', { event });
-    },
-    nip44: {
-      encrypt(pubkey, plaintext) {
-        return request('nip44.encrypt', { pubkey, plaintext });
-      },
-      decrypt(pubkey, ciphertext) {
-        return request('nip44.decrypt', { pubkey, ciphertext });
-      },
-    },
-  };
-
-  window.__divineNostrBridgeInstalled = true;
-
-  // Auto-login: seed localStorage so the app recognises the session.
-  try { $loginJs } catch (_) {}
-
-  // Signal that a NIP-07 signer is available.
-  window.dispatchEvent(new Event('nostr:ready'));
-  document.dispatchEvent(new CustomEvent('nlAuth', {
-    detail: { type: 'login', method: 'extension' },
-  }));
-})();
-''';
-}
-
-/// Normalises an app's allowed origins to `scheme://host[:port]` rules for
-/// `WebViewCompat.addWebMessageListener` on Android. Only `http`/`https`
-/// entries survive: `Uri.origin` is defined solely for those schemes and
-/// throws a `StateError` for anything else (`ws`/`wss`, `nostrsigner:`, …),
-/// and a websocket URL can never be a web frame's origin so it is meaningless
-/// for `addWebMessageListener` regardless. Unparseable entries and those with
-/// an empty host are dropped too, so the result is always a valid rule set the
-/// native side can pass through.
-@visibleForTesting
-List<String> webMessageAllowedOriginRules(List<String> allowedOrigins) {
-  const originSchemes = {'http', 'https'};
-  final rules = <String>[];
-  for (final origin in allowedOrigins) {
-    final parsed = Uri.tryParse(origin);
-    if (parsed == null ||
-        parsed.host.isEmpty ||
-        !originSchemes.contains(parsed.scheme)) {
-      continue;
-    }
-    rules.add(parsed.origin);
-  }
-  return rules;
-}
-
-/// Escapes a string for safe embedding inside a JS single-quoted
-/// literal.
-String _escapeJs(String value) {
-  return value
-      .replaceAll(r'\', r'\\')
-      .replaceAll("'", r"\'")
-      .replaceAll('`', r'\`')
-      .replaceAll('\n', r'\n')
-      .replaceAll('\r', r'\r');
-}
-
 class _SandboxStatusCard extends StatelessWidget {
   const _SandboxStatusCard({
     required this.title,
@@ -868,34 +715,4 @@ class _BootstrappedSandboxPage {
 
   final Uri baseUri;
   final String html;
-}
-
-@visibleForTesting
-String injectBridgeBootstrapIntoHtml(
-  String html, {
-  required String nonce,
-  String? pubkey,
-  String? autoLoginScript,
-}) {
-  final script = buildBridgeBootstrapScript(
-    nonce: nonce,
-    pubkey: pubkey,
-    autoLoginScript: autoLoginScript,
-  );
-  final bridgeMarkup = '<!-- divine-nostr-bridge --><script>$script</script>';
-  final insertionPoints = <RegExp>[
-    RegExp('<head[^>]*>', caseSensitive: false),
-    RegExp('<!doctype[^>]*>', caseSensitive: false),
-    RegExp('<html[^>]*>', caseSensitive: false),
-    RegExp('<body[^>]*>', caseSensitive: false),
-  ];
-
-  for (final insertionPoint in insertionPoints) {
-    final match = insertionPoint.firstMatch(html);
-    if (match != null) {
-      return html.replaceRange(match.end, match.end, bridgeMarkup);
-    }
-  }
-
-  return '$bridgeMarkup$html';
 }
